+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+Version 3.6.0 (build 2019060100)
+* Added caching to speed up overviews
+----------
+
+Version 3.5.0 (build 2018050300)
+* Added privacy provider
+* Validated functionality for Moodle 3.5 onwards
+* Minimum required Moodle version: 3.5
+----------
+
 Version 3.3.0 (build 2017092500)
 * Fixed deprecated pix_url references (replaced y image_url)
 * Added setting for maximum number of users to display on restore.php / restore form